+++ AppPackages/Sources/WorldOfPaybackAppModels/Transaction/TransactionDetail.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import WorldOfPaybackAppExtensions
 
 public struct TransactionDetail: Codable, Equatable, Hashable {
     public var description: String?
@@ -30,3 +31,15 @@ extension TransactionDetail {
         case value
     }
 }
+
+#if DEBUG
+public extension TransactionDetail {
+    static func sut() -> Self {
+        return .init(
+            description: "description",
+            bookingDate: DateFormatter.customDateFormatter.date(from: "2023-12-26T11:42:00"),
+            value: TransactionValue.sut
+        )
+    }
+}
+#endif
