+++ AppPackages/Tests/WorldOfPaybackAppComponentsTests/UIComponents/MainToolbarStoreTests.swift
@@ -5,9 +5,9 @@
 //  Created by Tsimur Asayonak on 12/26/23.
 //
 
-@testable import WorldOfPayback_Dev
 import ComposableArchitecture
 import XCTest
+import WorldOfPaybackAppComponents
 
 @MainActor
 final class MainToolbarStoreTests: XCTestCase {
